[Android] Don't filter modules in file browser
Let the file browser be a file browser and list all files in directory
instead of scanning valid module files. This should address problems
with people having many modules in a given directory. Modules are filtered
when added to a playlist or a play queue, invalid modules are not added.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>