Start support for multiple entry points
A module could have different pattern chains with different sub-songs or
"hidden" patterns that are not pat of the main song. The idea here is to
scan for all possible chains inside the module and publish the possible
"entry points" in the module structure. A front-end such as the Android
player could list them and allow the user to pick a different sub-song.

The default song is always present and always start at order 0.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>