[Android] Don't show invalid pattern on skip
We pass a delayed pattern and row information to patternViewer but it
loads current mod data. When skipping mods it shows current position
of the new pattern data.

We can't place pattern data in the latency compensation queue because
it would use too much space, so we reset the latency queue on each new
loaded module.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>