[Android] Fix playlist module info iteration
The same description was being added to all files when adding files to a
playlist. This was caused by a broken iteration over the file list. This
patch by Sage Herron fixes the issue.

The bug is described below:

from:	 sage <sage@54.org>
to:	 cmatsuoka@gmail.com
date:	 Sat, May 12, 2012 at 9:41 PM
subject: bug - when adding files to playlist the description is the same
	 for all of them

I think I've found a bug when adding files to a playlist, here are the
steps to re-create:

1) create a new playlist
2) long press on a directory which contains MOD files and click add to
play list and select a new playlist and touch OK
3) go back to the list of playlists and select the playlist
4) the name for each song is the name of the last file in the
directory which you added the playlist.
5) even though each song is not named correctly, it does play all the
files correctly that were added.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>