[Android] Fix oscilloscope update
Oscilloscope waveform is to be reset at every new note event. However,
viewer update is asynchronous to play events and the current key event
is active during all frames of row (can't set it during first row only
because it may be missed by the asynchronous viewer poll). Change key
information retrieval in such a way that key event is retrieved only
once per row. This will fail in case of one-row patterns, but this is
a very rare situation.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>